move class MatchRef to genotype-order.js in preparation for adding null
move class MatchRef from manage-genotype.js : filterSamples() to genotype-order.js. This places it adjacent to MatchRefSample which has the same signature, and localises the Alt/Ref/null design change in genotype-order.js, and reduces the size of manage-genotype.js.
(origin : in cafd7623 MatchRef was factored from haplotypeFilterSamples(), later renamed to filterSamples()).
